logging: Redesign how the convenience wrappers work
Considering the philosophy of the logging api is brevity and to stay out of the way, after a chat with Kojin I've had a go at rethinking the way the api should function.
Kojin remarked that categories tend to be per-file. Instead of generating the convenience wrappers once in logger.hpp, none are generated globally and instead each file can create theirs with a given category instead of having to specify the category for each log call.
For any sort of logging requirements that breaks this mould the full unwrapped functions are available and their signatures have been improved for clarity.
The logging tester was also updated to test & demonstrate both methods of using the logging system.

Oh yeah I fixed a typo too.